requirements.txt
asgiref==3.4.1
autopep8==1.5.7
coverage==5.5
decouple==0.0.7
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
drf-nested-routers==0.93.3
gunicorn==20.1.0
psycopg2==2.9.1
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
